refactor: 💡 move non-db related functions in to own file
